--- a/benches/serde.rs
+++ b/benches/serde.rs
@@ -3,6 +3,7 @@
 extern crate test;
 
 extern crate avro_rs;
+extern crate serde_json;
 use avro_rs::schema::Schema;
 use avro_rs::types::{Record, ToAvro, Value};
 use avro_rs::Reader;
@@ -152,8 +153,33 @@ fn make_big_record() -> (Schema, Value) {
     (big_schema, big_record)
 }
 
+fn make_big_json_record() -> serde_json::Value {
+    let mut address = std::collections::HashMap::new();
+    address.insert("street".to_owned(), "street".to_owned());
+    address.insert("city".to_owned(), "city".to_owned());
+    address.insert("state_prov".to_owned(), "state_prov".to_owned());
+    address.insert("country".to_owned(), "country".to_owned());
+    address.insert("zip".to_owned(), "zip".to_owned());
+    let address_json = serde_json::to_value(address).unwrap();
+    let mut big_record = std::collections::HashMap::new();
+    big_record.insert("username".to_owned(), serde_json::to_value("username").unwrap());
+    big_record.insert("age".to_owned(), serde_json::to_value(10i32).unwrap());
+    big_record.insert("phone".to_owned(), serde_json::to_value("000000000").unwrap());
+    big_record.insert("housenum".to_owned(), serde_json::to_value("0000").unwrap());
+    big_record.insert("address".to_owned(), address_json);
+    serde_json::to_value(big_record).unwrap()
+}
+
 fn make_records(record: &Value, count: usize) -> Vec<Value> {
     let mut records = Vec::new();
+    for _ in 0..count {
+        records.push(record.clone());
+    }
+    records
+}
+
+fn make_json_records(record: &serde_json::Value, count: usize) -> Vec<serde_json::Value> {
+    let mut records = Vec::with_capacity(count);
     for _ in 0..count {
         records.push(record.clone());
     }
@@ -166,9 +192,20 @@ fn write(schema: &Schema, records: &[Value]) -> Vec<u8> {
     writer.into_inner()
 }
 
+fn write_json(records: &[serde_json::Value]) -> Vec<u8> {
+    serde_json::to_vec(records).unwrap()
+}
+
 fn read(schema: &Schema, bytes: &[u8]) {
     let reader = Reader::with_schema(schema, bytes).unwrap();
     for record in reader {
+        let _ = record;
+    }
+}
+
+fn read_json(bytes: &[u8]) {
+    let reader: serde_json::Value = serde_json::from_slice(bytes).unwrap();
+    for record in reader.as_array().unwrap() {
         let _ = record;
     }
 }
@@ -186,6 +223,15 @@ fn bench_read(b: &mut test::Bencher, make_record: &Fn() -> (Schema, Value), n_re
     println!("bytes.len() = {}", bytes.len());
     println!("records.len() = {}", records.len());
     b.iter(|| read(&schema, &bytes));
+}
+
+fn bench_read_json(b: &mut test::Bencher, make_record: &Fn() -> (serde_json::Value), n_records: usize) {
+    let record = make_record();
+    let records = make_json_records(&record, n_records);
+    let bytes = write_json(&records);
+    println!("bytes.len() = {}", bytes.len());
+    println!("records.len() = {}", records.len());
+    b.iter(|| read_json(&bytes));
 }
 
 #[bench]
@@ -251,4 +297,9 @@ fn bench_big_schema_read_10000_record(b: &mut test::Bencher) {
 #[bench]
 fn bench_big_schema_read_100000_record(b: &mut test::Bencher) {
     bench_read(b, &make_big_record, 100000);
+}
+
+#[bench]
+fn bench_big_schema_json_read_100000_record(b: &mut test::Bencher) {
+    bench_read_json(b, &make_big_json_record, 100000);
 }