--- a/src/decode.rs
+++ b/src/decode.rs
@@ -31,35 +31,39 @@ pub fn decode<R: Read>(schema: &Schema, reader: &mut R) -> Result<Value, Error> 
                 1u8 => Ok(Value::Boolean(true)),
                 _ => Err(DecodeError::new("not a bool").into()),
             }
-        },
+        }
         &Schema::Int => decode_int(reader),
         &Schema::Long => decode_long(reader),
         &Schema::Float => {
             let mut buf = [0u8; 4];
             reader.read_exact(&mut buf[..])?;
             Ok(Value::Float(unsafe { transmute::<[u8; 4], f32>(buf) }))
-        },
+        }
         &Schema::Double => {
             let mut buf = [0u8; 8];
             reader.read_exact(&mut buf[..])?;
             Ok(Value::Double(unsafe { transmute::<[u8; 8], f64>(buf) }))
-        },
+        }
         &Schema::Bytes => {
             if let Value::Long(len) = decode_long(reader)? {
                 let len = len as usize;
                 let mut buf = Vec::with_capacity(len);
-                unsafe { buf.set_len(len); }
+                unsafe {
+                    buf.set_len(len);
+                }
                 reader.read_exact(&mut buf)?;
                 Ok(Value::Bytes(buf))
             } else {
                 Err(DecodeError::new("bytes len not found").into())
             }
-        },
+        }
         &Schema::String => {
             if let Value::Long(len) = decode_long(reader)? {
                 let len = len as usize;
                 let mut buf = Vec::with_capacity(len);
-                unsafe { buf.set_len(len); }
+                unsafe {
+                    buf.set_len(len);
+                }
                 reader.read_exact(&mut buf)?;
 
                 String::from_utf8(buf)
@@ -68,12 +72,12 @@ pub fn decode<R: Read>(schema: &Schema, reader: &mut R) -> Result<Value, Error> 
             } else {
                 Err(DecodeError::new("string len not found").into())
             }
-        },
+        }
         &Schema::Fixed { size, .. } => {
             let mut buf = vec![0u8; size as usize];
             reader.read_exact(&mut buf)?;
             Ok(Value::Fixed(size, buf))
-        },
+        }
         &Schema::Array(ref inner) => {
             let mut items = Vec::new();
 
@@ -82,7 +86,7 @@ pub fn decode<R: Read>(schema: &Schema, reader: &mut R) -> Result<Value, Error> 
                     // arrays are 0-terminated, 0i64 is also encoded as 0 in Avro
                     // reading a length of 0 means the end of the array
                     if len == 0 {
-                        break
+                        break;
                     }
 
                     items.reserve(len as usize);
@@ -90,12 +94,12 @@ pub fn decode<R: Read>(schema: &Schema, reader: &mut R) -> Result<Value, Error> 
                         items.push(decode(inner, reader)?);
                     }
                 } else {
-                    return Err(DecodeError::new("array len not found").into())
+                    return Err(DecodeError::new("array len not found").into());
                 }
             }
 
             Ok(Value::Array(items))
-        },
+        }
         &Schema::Map(ref inner) => {
             let mut items = HashMap::new();
 
@@ -104,7 +108,7 @@ pub fn decode<R: Read>(schema: &Schema, reader: &mut R) -> Result<Value, Error> 
                     // maps are 0-terminated, 0i64 is also encoded as 0 in Avro
                     // reading a length of 0 means the end of the map
                     if len == 0 {
-                        break
+                        break;
                     }
 
                     items.reserve(len as usize);
@@ -113,16 +117,16 @@ pub fn decode<R: Read>(schema: &Schema, reader: &mut R) -> Result<Value, Error> 
                             let value = decode(inner, reader)?;
                             items.insert(key, value);
                         } else {
-                            return Err(DecodeError::new("map key is not a string").into())
+                            return Err(DecodeError::new("map key is not a string").into());
                         }
                     }
                 } else {
-                    return Err(DecodeError::new("map len not found").into())
+                    return Err(DecodeError::new("map len not found").into());
                 }
             }
 
             Ok(Value::Map(items))
-        },
+        }
         &Schema::Union(ref inner) => {
             let index = zag_i64(reader)?;
 
@@ -131,25 +135,21 @@ pub fn decode<R: Read>(schema: &Schema, reader: &mut R) -> Result<Value, Error> 
                 1 => decode(inner, reader).map(|x| Value::Union(Some(Box::new(x)))),
                 _ => Err(DecodeError::new("union index out of bounds").into()),
             }
-        },
+        }
         &Schema::Record { ref fields, .. } => {
             // Benchmarks indicate ~10% improvement using this method.
-            // Putting it in the `Value::Record` initially may be overkill, but
-            // not using the Iterator API appears to perform better.
-            let mut records = Value::Record(Vec::new());
-            if let &mut Value::Record(ref mut items) = &mut records {
-                for field in fields {
-                    // This clone is also expensive. See if we can do away with it...
-                    items.push((field.name.clone(), decode(&field.schema, reader)?));
-                }
+            let mut items = Vec::new();
+            for field in fields {
+                // This clone is also expensive. See if we can do away with it...
+                items.push((field.name.clone(), decode(&field.schema, reader)?));
             }
-            Ok(records)
+            Ok(Value::Record(items))
             // fields
             // .iter()
             // .map(|field| decode(&field.schema, reader).map(|value| (field.name.clone(), value)))
             // .collect::<Result<Vec<(String, Value)>, _>>()
             // .map(|items| Value::Record(items))
-        },
+        }
         &Schema::Enum { ref symbols, .. } => {
             if let Value::Int(index) = decode_int(reader)? {
                 if index >= 0 && (index as usize) <= symbols.len() {
@@ -161,6 +161,6 @@ pub fn decode<R: Read>(schema: &Schema, reader: &mut R) -> Result<Value, Error> 
             } else {
                 Err(DecodeError::new("enum symbol not found").into())
             }
-        },
+        }
     }
 }