… benches/quickstop-null.avro
… src/schema.rs
… src/writer.rs