~~**Not yet ready to merge, need to clean up some and add more comments to new functions.**~~

I did some profiling with `perf` on the benchmarks. I found that `clone()`ing the input `Vec`s was adding a lot of overhead, so I changed the benchmarks to accept slices instead. I also found that the `read` benchmark wasn't `unwrap()`ing the `Result<Reader, _>`, so it wasn't actually running.

Please let me know if you have any objections, concerns, questions, etc.!

Here is a high level list of changes made:
### Read
* Check if the `write_schema` and `read_schema` are the same. If so, don't run schema resolution. This improved read time by `~50%`.
* Added an internal `Block<R>` struct that can read byte blocks into an internal buffer.
  This reduces the number of allocations that occur. `Block<R>` then incrementally decodes `Value` objects, instead of storing them in a `VecDeque<Value>`.
  This further improved read time by `~30%`.
* Reduced recursive calls, allocations and moves in `decode`. `~10%` improvement.

### Write
* `encode` was spending a lot of time moving data around. I added an `encode_ref` function which operates on a `&Value`. This improved performance as well (will provide numbers soon).
* Added pub fn `extend_from_slice(&mut self, slice: &[Value])`.

### Other Benchmark Changes
* Added a comparable JSON benchmark which uses `serde_json` for comparison.
* Added a `bench_from_file` function and a file from the `goavro` repository for comparison.

## Results

These tests were compiled with the latest `rust nightly` use the command:
```
cargo +nightly bench -- read
```
These were all captured on my local machine, which was mostly idle other than running the tests.
Running `Intel(R) Core(TM) i7-7820HQ CPU @ 2.90GHz` on Ubuntu 16.04.

Initial benchmark:
```
test bench_big_schema_read_100000_record   ... bench: 341,941,981 ns/iter (+/- 27,649,636)
test bench_big_schema_read_10000_record    ... bench:  34,915,152 ns/iter (+/- 3,895,246)
test bench_big_schema_read_100_record      ... bench:     362,215 ns/iter (+/- 42,285)
test bench_big_schema_read_1_record        ... bench:      14,903 ns/iter (+/- 1,039)
test bench_small_schema_read_10000_record  ... bench:   4,694,758 ns/iter (+/- 435,888)
test bench_small_schema_read_100_record    ... bench:      45,600 ns/iter (+/- 2,255)
test bench_small_schema_read_1_record      ... bench:       2,984 ns/iter (+/- 251)
```

After these changes:
```
test bench_big_schema_read_100000_record      ... bench: 102,894,070 ns/iter (+/- 1,688,040)
test bench_big_schema_read_10000_record       ... bench:  10,228,593 ns/iter (+/- 339,369)
test bench_big_schema_read_100_record         ... bench:     115,083 ns/iter (+/- 4,408)
test bench_big_schema_read_1_record           ... bench:      13,388 ns/iter (+/- 693)
test bench_small_schema_read_10000_record     ... bench:   1,466,798 ns/iter (+/- 37,203)
test bench_small_schema_read_100_record       ... bench:      17,075 ns/iter (+/- 758)
test bench_small_schema_read_1_record         ... bench:       2,658 ns/iter (+/- 342)
```

Comparison with `serde_json`:
```
test bench_big_schema_json_read_100000_record ... bench: 260,100,006 ns/iter (+/- 4,294,564)
test bench_big_schema_read_100000_record      ... bench: 102,894,070 ns/iter (+/- 1,688,040)
```