@Snnappie 2018-06-14
@poros @flavray Thanks for reviewing all these changes! -- In the future if I have more contributions I'll try to make the pull request smaller in scope so the review isn't as arduous.

I'll address your latest comments, and I'd like to document the methods added to `Writer` to the same standard as the others before I remove the `WIP` status.